Move discount related fixtures from other files
