+++ saleor/discount/tests/fixtures/__init__.py
@@ -1,3 +1,4 @@
-from .voucher import *  # noqa: F403
+from .benchmark import *  # noqa: F403
 from .promotion import *  # noqa: F403
-from .manual_discount import *  # noqa: F403
+from .promotion_rule import *  # noqa: F403
+from .voucher import *  # noqa: F403
