+++ saleor/discount/tests/fixtures/promotion.py
@@ -1,115 +1,14 @@
 import datetime
-import uuid
-from contextlib import contextmanager
 from decimal import Decimal
-from functools import partial
-from io import BytesIO
-from typing import Callable, Optional
-from unittest.mock import MagicMock
 
 import graphene
 import pytest
-from django.conf import settings
-from django.contrib.sites.models import Site
-from django.core.files.uploadedfile import SimpleUploadedFile
-from django.db import connection
-from django.template.defaultfilters import truncatechars
-from django.test.utils import CaptureQueriesContext as BaseCaptureQueriesContext
 from django.utils import timezone
-from django_countries import countries
-from freezegun import freeze_time
-from PIL import Image
-from prices import Money
 
-from saleor.account.models import Address, Group, StaffNotificationRecipient, User
-from saleor.attribute import AttributeEntityType, AttributeInputType, AttributeType
-from saleor.attribute.models import (
-    Attribute,
-    AttributeTranslation,
-    AttributeValue,
-    AttributeValueTranslation,
-)
-from saleor.attribute.utils import associate_attribute_values_to_instance
-from saleor.checkout import base_calculations
-from saleor.checkout.fetch import fetch_checkout_info, fetch_checkout_lines
-from saleor.checkout.models import Checkout, CheckoutLine, CheckoutMetadata
-from saleor.checkout.utils import (
-    add_variant_to_checkout,
-    add_voucher_to_checkout,
-)
-from saleor.core import JobStatus
-from saleor.core.models import EventDelivery, EventDeliveryAttempt, EventPayload
-from saleor.core.payments import PaymentInterface
-from saleor.core.units import MeasurementUnits
-from saleor.core.utils.editorjs import clean_editor_js
-from saleor.csv.events import ExportEvents
-from saleor.csv.models import ExportEvent, ExportFile
-from saleor.discount import (
-    DiscountType,
-    DiscountValueType,
-    PromotionEvents,
-    PromotionType,
-    RewardType,
-    RewardValueType,
-    VoucherType,
-)
-from saleor.discount.interface import VariantPromotionRuleInfo
-from saleor.discount.models import (
-    CheckoutDiscount,
-    CheckoutLineDiscount,
-    Promotion,
-    PromotionEvent,
-    PromotionRule,
-    PromotionRuleTranslation,
-    PromotionTranslation,
-    Voucher,
-    VoucherChannelListing,
-    VoucherCode,
-    VoucherCustomer,
-    VoucherTranslation,
-)
-from saleor.giftcard import GiftCardEvents
-from saleor.giftcard.models import GiftCard, GiftCardEvent, GiftCardTag
-from saleor.menu.models import Menu
-from saleor.payment import ChargeStatus, TransactionKind
-from saleor.payment.interface import (
-    AddressData,
-    GatewayConfig,
-    GatewayResponse,
-    PaymentData,
-)
-from saleor.payment.models import Payment, TransactionEvent, TransactionItem
-from saleor.payment.transaction_item_calculations import recalculate_transaction_amounts
-from saleor.payment.utils import create_manual_adjustment_events
-from saleor.permission.enums import get_permissions
-from saleor.permission.models import Permission
-from saleor.plugins.manager import get_plugins_manager
-from saleor.product.models import (
-    CategoryTranslation,
-    CollectionTranslation,
-    ProductTranslation,
-    ProductVariantChannelListing,
-    ProductVariantTranslation,
-)
-from saleor.product.utils.variants import fetch_variants_for_promotion_rules
-from saleor.shipping.models import (
-    ShippingMethod,
-    ShippingMethodChannelListing,
-    ShippingMethodTranslation,
-    ShippingMethodType,
-    ShippingZone,
-)
-from saleor.shipping.utils import convert_to_shipping_method_data
-from saleor.site.models import SiteSettings
-from saleor.tax import TaxCalculationStrategy
-from saleor.warehouse.models import (
-    PreorderReservation,
-    Reservation,
-    Warehouse,
-)
-from saleor.webhook.event_types import WebhookEventAsyncType
-from saleor.webhook.transport.utils import to_payment_app_id
-from saleor.tests.utils import dummy_editorjs
+from ....product.utils.variants import fetch_variants_for_promotion_rules
+from ....tests.utils import dummy_editorjs
+from ... import PromotionType, RewardType, RewardValueType
+from ...models import Promotion, PromotionRule
 
 
 @pytest.fixture
@@ -310,6 +209,29 @@ def promotion_list(channel_USD, product, collection):
 
 
 @pytest.fixture
+def promotion_10_percentage(channel_USD, product_list, product):
+    promotion = Promotion.objects.create(
+        name="Promotion",
+    )
+    product_list.append(product)
+    rule = promotion.rules.create(
+        name="10% promotion rule",
+        catalogue_predicate={
+            "productPredicate": {
+                "ids": [
+                    graphene.Node.to_global_id("Product", product.id)
+                    for product in product_list
+                ]
+            }
+        },
+        reward_value_type=RewardValueType.PERCENTAGE,
+        reward_value=Decimal("10"),
+    )
+    rule.channels.add(channel_USD)
+    return promotion
+
+
+@pytest.fixture
 def promotion_converted_from_sale(catalogue_predicate, channel_USD):
     promotion = Promotion.objects.create(name="Sale", type=PromotionType.CATALOGUE)
     promotion.assign_old_sale_id()
@@ -361,3 +283,53 @@ def promotion_converted_from_sale_with_empty_predicate(channel_USD):
     )
     rule.channels.add(channel_USD)
     return promotion
+
+
+@pytest.fixture
+def promotion_converted_from_sale_list(channel_USD, product_list, category, collection):
+    promotions = Promotion.objects.bulk_create(
+        [Promotion(name="Sale 1"), Promotion(name="Sale 2"), Promotion(name="Sale 3")]
+    )
+    for promotion in promotions:
+        promotion.assign_old_sale_id()
+    predicates = [
+        {
+            "OR": [
+                {
+                    "productPredicate": {
+                        "ids": [graphene.Node.to_global_id("Product", product.id)]
+                    }
+                },
+                {
+                    "variantPredicate": {
+                        "ids": [
+                            graphene.Node.to_global_id(
+                                "Product", product.variants.first().id
+                            )
+                        ]
+                    }
+                },
+            ]
+        }
+        for product in product_list
+    ]
+    collection_id = graphene.Node.to_global_id("Collection", collection.id)
+    category_id = graphene.Node.to_global_id("Category", category.id)
+
+    predicates[0]["OR"].append({"categoryPredicate": {"ids": [category_id]}})
+    predicates[1]["OR"].append({"collectionPredicate": {"ids": [collection_id]}})
+
+    rules = [
+        PromotionRule(
+            promotion=promotion,
+            catalogue_predicate=predicate,
+            reward_value_type=RewardValueType.FIXED,
+            reward_value=Decimal("5"),
+        )
+        for promotion, predicate in zip(promotions, predicates)
+    ]
+    PromotionRule.objects.bulk_create(rules)
+    for rule in rules:
+        rule.channels.add(channel_USD)
+
+    return promotions
