+++ saleor/discount/tests/test_rounding_issue.py
@@ -1,6 +1,5 @@
 from decimal import Decimal
 
-import graphene
 import pytest
 
 from ...checkout.fetch import fetch_checkout_info, fetch_checkout_lines
@@ -10,34 +9,10 @@ from ...plugins.manager import get_plugins_manager
 from ...product.models import Product, ProductVariantChannelListing
 from ...product.utils.variant_prices import update_discounted_prices_for_promotion
 from ...product.utils.variants import fetch_variants_for_promotion_rules
-from .. import DiscountValueType, RewardValueType
-from ..models import Promotion
+from .. import DiscountValueType
 from ..utils.checkout import (
     create_or_update_discount_objects_from_promotion_for_checkout,
 )
-
-
-@pytest.fixture
-def promotion_10_percentage(channel_USD, product_list, product):
-    promotion = Promotion.objects.create(
-        name="Promotion",
-    )
-    product_list.append(product)
-    rule = promotion.rules.create(
-        name="10% promotion rule",
-        catalogue_predicate={
-            "productPredicate": {
-                "ids": [
-                    graphene.Node.to_global_id("Product", product.id)
-                    for product in product_list
-                ]
-            }
-        },
-        reward_value_type=RewardValueType.PERCENTAGE,
-        reward_value=Decimal("10"),
-    )
-    rule.channels.add(channel_USD)
-    return promotion
 
 
 @pytest.mark.parametrize(
