+++ saleor/graphql/discount/tests/benchmark/test_sales.py
@@ -1,51 +1,6 @@
 import pytest
 
-from .....discount import RewardValueType
-from .....discount.models import Promotion, PromotionRule
 from ....tests.utils import get_graphql_content
-
-
-@pytest.fixture
-def promotion_converted_from_sale_list(channel_USD, channel_PLN):
-    promotions = Promotion.objects.bulk_create(
-        [Promotion(name="Sale1"), Promotion(name="Sale2"), Promotion(name="Sale2")]
-    )
-    for promotion in promotions:
-        promotion.assign_old_sale_id()
-
-    values = [15, 5, 25]
-    usd_rules, pln_rules = [], []
-    for promotion, value in zip(promotions, values):
-        usd_rules.append(
-            PromotionRule(
-                promotion=promotion,
-                catalogue_predicate={},
-                reward_value_type=RewardValueType.FIXED,
-                reward_value=value,
-            )
-        )
-        pln_rules.append(
-            PromotionRule(
-                promotion=promotion,
-                catalogue_predicate={},
-                reward_value_type=RewardValueType.FIXED,
-                reward_value=value * 2,
-            )
-        )
-    PromotionRule.objects.bulk_create(usd_rules + pln_rules)
-    PromotionRuleChannel = PromotionRule.channels.through
-    usd_rules_channels = [
-        PromotionRuleChannel(promotionrule_id=rule.id, channel_id=channel_USD.id)
-        for rule in usd_rules
-    ]
-    pln_rules_channels = [
-        PromotionRuleChannel(promotionrule_id=rule.id, channel_id=channel_PLN.id)
-        for rule in usd_rules
-    ]
-    PromotionRuleChannel.objects.bulk_create(usd_rules_channels + pln_rules_channels)
-
-    return promotions
-
 
 SALES_QUERY = """
 query GetSales($channel: String){
@@ -110,7 +65,7 @@ query GetSales($channel: String){
 @pytest.mark.count_queries(autouse=False)
 def test_sales_query_with_channel_slug(
     staff_api_client,
-    promotion_converted_from_sale_list,
+    promotion_converted_from_sale_list_for_benchmark,
     channel_USD,
     permission_manage_discounts,
     count_queries,
@@ -130,7 +85,7 @@ def test_sales_query_with_channel_slug(
 @pytest.mark.count_queries(autouse=False)
 def test_sales_query_without_channel_slug(
     staff_api_client,
-    promotion_converted_from_sale_list,
+    promotion_converted_from_sale_list_for_benchmark,
     permission_manage_discounts,
     count_queries,
 ):
