+++ saleor/graphql/discount/tests/mutations/test_voucher_bulk_delete.py
@@ -1,53 +1,9 @@
 from unittest import mock
 
 import graphene
-import pytest
 
-from .....discount.models import Voucher, VoucherChannelListing, VoucherCode
+from .....discount.models import Voucher
 from ....tests.utils import get_graphql_content
-
-
-@pytest.fixture
-def voucher_list(channel_USD):
-    [voucher_1, voucher_2, voucher_3] = Voucher.objects.bulk_create(
-        [
-            Voucher(),
-            Voucher(),
-            Voucher(),
-        ]
-    )
-
-    VoucherCode.objects.bulk_create(
-        [
-            VoucherCode(code="voucher-1", voucher=voucher_1),
-            VoucherCode(code="voucher-2", voucher=voucher_1),
-            VoucherCode(code="voucher-3", voucher=voucher_2),
-        ]
-    )
-    VoucherChannelListing.objects.bulk_create(
-        [
-            VoucherChannelListing(
-                voucher=voucher_1,
-                channel=channel_USD,
-                discount_value=1,
-                currency=channel_USD.currency_code,
-            ),
-            VoucherChannelListing(
-                voucher=voucher_2,
-                channel=channel_USD,
-                discount_value=2,
-                currency=channel_USD.currency_code,
-            ),
-            VoucherChannelListing(
-                voucher=voucher_3,
-                channel=channel_USD,
-                discount_value=3,
-                currency=channel_USD.currency_code,
-            ),
-        ]
-    )
-    return voucher_1, voucher_2, voucher_3
-
 
 BULK_DELETE_VOUCHERS_MUTATION = """
     mutation voucherBulkDelete($ids: [ID!]!) {
